trading_bot_package/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
streamlit>=1.10.0
alpaca-trade-api>=2.1.0
python-dotenv>=0.19.0
yfinance>=0.1.70
requests>=2.27.0
schedule>=1.1.0
plotly>=5.5.0
